Split the files
Added more files and split them so it looks better and that theres less code in landing.js
