+++ frontend/web/src/styles/sass/pages/landingpage/components/_contactForm.scss
@@ -1,0 +1,219 @@
+.questionsContainer {
+  width: 100%;
+  margin-top: 50px;
+  background-color: #f9f9f9;
+  padding: 20px;
+  box-sizing: border-box;
+
+  h1,
+  p {
+    margin: 15px 0;
+  }
+
+  .contact {
+    width: 95%;
+    margin: 50px auto;
+    font-weight: 500;
+    color: #1a1a1a;
+    font-size: 1.1em;
+    margin-bottom: 50px;
+  }
+
+  .contactform {
+    display: flex;
+    flex-wrap: wrap;
+    margin-top: 25px;
+
+    input:not(.biginput input) {
+      border-radius: 10px;
+      background: #FFF;
+      padding: 10px;
+      width: 100%;
+      margin: 15px 0 10px 0;
+      border: 1px solid #ddd;
+      box-sizing: border-box;
+
+      &::placeholder {
+        text-align: left;
+      }
+    }
+
+    textarea::placeholder {
+      text-align: left;
+    }
+  }
+
+  .inputContainer {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+
+    label {
+      flex: 1 1 45%;
+      display: block;
+      font-weight: 500;
+      margin-bottom: 5px;
+      color: #333;
+
+      input,
+      textarea {
+        width: 100%;
+        padding: 10px;
+        border-radius: 5px;
+        border: 1px solid #ddd;
+        box-sizing: border-box;
+        margin-bottom: 15px;
+
+        &::placeholder {
+          color: #666;
+        }
+      }
+    }
+  }
+
+  .biginput {
+    label {
+      display: block;
+      font-weight: 500;
+      margin-bottom: 5px;
+      color: #333;
+    }
+  }
+
+  .biginput textarea {
+    width: 100%;
+    height: 240px;
+    margin-top: 10px;
+    box-sizing: border-box;
+    text-align: left;
+    padding: 10px;
+    border-radius: 10px;
+    border: 1px solid #ddd;
+    resize: vertical;
+    font-family: inherit;
+  }
+
+  .biginput {
+    label {
+      display: block;
+      font-weight: 500;
+      margin-bottom: 5px;
+      color: #333;
+    }
+  }
+
+  .formbuttons {
+    border: 0;
+    display: flex;
+    justify-content: flex-start;
+    flex-wrap: wrap;
+    width: 100%;
+    margin-top: 20px;
+  }
+
+  #attachmentsbutton,
+  #sendbutton {
+    width: 47%;
+    height: 40px;
+    padding: 8px 15px;
+    margin-bottom: 20px;
+    flex-shrink: 0;
+    color: #FFF;
+    border-radius: 10px;
+    border: none;
+    box-shadow: none;
+    font-size: 16px;
+    font-weight: 500;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
+
+  #attachmentsbutton {
+    background: #003366;
+
+    &:hover {
+      background: #3e85cb;
+      cursor: pointer;
+    }
+  }
+
+  #sendbutton {
+    background: #0D9813;
+
+    &:hover {
+      background: #46c54c;
+      cursor: pointer;
+    }
+  }
+
+  .error-message {
+    color: #d32f2f;
+    font-size: 12px;
+    margin-top: 5px;
+    display: block;
+    text-align: left;
+  }
+
+  .feedback {
+    padding: 10px;
+    border-radius: 5px;
+    margin: 15px 0;
+    text-align: center;
+    font-weight: 500;
+
+    &.success {
+      background-color: #d4edda;
+      color: #155724;
+      border: 1px solid #c3e6cb;
+    }
+
+    &.error {
+      background-color: #f8d7da;
+      color: #721c24;
+      border: 1px solid #f5c6cb;
+    }
+  }
+}
+
+@media screen and (max-width: 930px) {
+  .questionsContainer {
+    .contact {
+      padding: 10px;
+      margin: 10px auto;
+      font-size: 14px;
+    }
+
+    .contactform {
+      padding: 10px;
+    }
+
+    .namemessage,
+    .biginput,
+    .formbuttons {
+      flex: 100%;
+      width: 100%;
+      margin-right: 0;
+    }
+
+    .biginput textarea {
+      height: 100px;
+      padding: 8px;
+      width: 100%;
+    }
+
+    .formbuttons {
+      flex-direction: column;
+      align-items: center;
+
+      #attachmentsbutton,
+      #sendbutton {
+        width: 100%;
+        max-width: 300px;
+        margin-left: 0;
+        margin-bottom: 15px;
+      }
+    }
+  }
+}
+
+
